polykybd: rename corne42 variant to split42
Rename the 42-key variant from corne42 to split42 for naming consistency with
split72 (both are PolyKybd "split" boards). Same hardware, same 42-key CRKBD
footprint (LAYOUT_crkbd is kept — it names the physical layout), same USB PID
(0x2008).

  * directory  keyboards/handwired/polykybd/corne42/ -> split42/
  * files      corne42.c/.h -> split42.c/.h
  * build target handwired/polykybd/corne42 -> handwired/polykybd/split42
  * keyboard_name "PolyKybd Corne42" -> "PolyKybd Split42"
  * polykybd.h variant include, CI workflow (qmk-test.yml), and in-tree
    code comments updated.

split72 remains byte-identical (verified: same section sizes, all 1251
functions identical). split42 builds clean and carries full feature parity
via the shared poly_keymap.c.
