+++ keyboards/handwired/polykybd/oled_helper.h
@@ -9,7 +9,7 @@
 void num_to_u32_string(char* buffer, uint8_t buffer_len, uint8_t value);
 void hex_to_u32_string(char* buffer, uint8_t buffer_len, uint8_t value);
 
-/* Board-specific callbacks — implemented in split72/status_oled.c or corne42/status_oled.c */
+/* Board-specific callbacks — implemented in split72/status_oled.c or split42/status_oled.c */
 void oled_update_buffer(void);
 void oled_draw_kybd(void);
 void oled_draw_poly(void);
