+++ keyboards/handwired/polykybd/polykybd.h
@@ -23,8 +23,7 @@
 #if defined(KEYBOARD_handwired_polykybd_split72)
 #    include "split72/split72.h"
 #elif defined(KEYBOARD_handwired_polykybd_split42)
+// split42 was formerly named "corne42" (renamed 2026-06; same 42-key hardware).
 #    include "split42/split42.h"
-#elif defined(KEYBOARD_handwired_polykybd_corne42)
-#    include "corne42/corne42.h"
 #endif
