+++ keyboards/handwired/polykybd/split42/halconf.h
