+++ keyboards/handwired/polykybd/split42/keyboard.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "PolyKybd Corne42",
+    "keyboard_name": "PolyKybd Split42",
     "manufacturer": "PolyFabriq",
     "url": "https://ko-fi.com/polykb",
     "maintainer": "[thpoll]",
