+++ keyboards/handwired/polykybd/split42/rules.mk
@@ -8,13 +8,13 @@ SPLIT_KEYBOARD = yes
 OLED_ENABLE = yes
 OLED_DRIVER = ssd1306
 
-# No RGB matrix — corne42 has no underglow or per-key LEDs
+# No RGB matrix — split42 has no underglow or per-key LEDs
 
-# Source files — status_oled.c resolves to corne42/status_oled.c via QMK search order
+# Source files — status_oled.c resolves to split42/status_oled.c via QMK search order
 QUANTUM_LIB_SRC += spi_master.c
 SRC += status_oled.c base/update.c base/e2prom.c base/com.c base/helpers.c base/disp_array.c base/shift_reg.c base/spi_helper.c base/overlay.c base/multicore/core1.c lang/lang_lut.c base/fw_staging.c
 
-# No pointing device (no Cirque trackpad on corne42)
+# No pointing device (no Cirque trackpad on split42)
 
 RAW_ENABLE     = yes
 WPM_ENABLE     = yes
